+++ src/admin/config/clientsImportExport.ts
@@ -1,4 +1,5 @@
 import type { FieldConfig } from '../types/import-export'
+import { parseDecimal } from '../types/import-export'
 
 export const CLIENTS_FIELDS_CONFIG: FieldConfig[] = [
   { key: 'id', label: 'ID', exportable: true, importable: true },
@@ -13,7 +14,7 @@ export const CLIENTS_FIELDS_CONFIG: FieldConfig[] = [
     exportable: true,
     importable: true,
     formatter: {
-      import: v => (v !== '' ? parseFloat(v) : null),
+      import: v => (v !== '' ? parseDecimal(v) : null),
     },
   },
   {
